Adding initial project
Using a preview version of the BAS Metadata Library to generate candidate records with XML stylesheets for visualisation and data quality.

This designed to help develop a candidate record from a minimal record to something as complete as feasible.

In time this project/repository will morph into tooling specific to Arc for importing/exporting records from a CSW catalogue.