Fixing enabling Sentry in production environments
It seems the str2bool function returns None instead of False if the input to that method is False (i.e. already a negative boolean value).

Casting the input to a string (becoming ‘False’) returns False (boolean) works as expected.

This feels like a bug to me but is probably logical.