Fixing item bounding boxes
Up to now we have always overridden the bounding box for items based on a st of well-known extents, which use far more complex geometries to around crossing the pole etc.

We now have items that define a item-specific extent and so need to use the values directly from the metadata record. This worked but we incorrectly used the 3031 CRS for all bounding boxes, rather than only overriding it where we use a well-known extent.

This is fixed.