Upgrading/integrating Metadata Library v0.8.0
This includes a series of changes that significantly increases the copying of PyCSW classes into the hazmat module.

These additions are needed to work around trailing XML tags, which should not be possible, and should not have changed as a result of this upgrade.

This will be addressed as part of a general review/refactor of our PyCSW dependency.