Refactoring collections to use records
Currently collections are bespoke objects with a local repository file. This file prevents the catalogue being managed in a distributed way.

This has been changed so that Collections are defined as records, the same as Items. This simplifies how resources are managed (everything is stored in a CSW repository) and removes the need for parallel, but essentially the same, commands.