Fixing tests following auth changes
Most tests are unchanged but where the mocked auth, large changes have been made to reduce what needs mocking.

This means requests made via Flask test (http) clients need to include (or not include) an access token rather than mocking the server side to fake this.

The biggest changes are in the auth commands, which are rewritten based on the client_auth internal extension, and to fixtures, which to avoid repetition are generated ny a factory/helper method.

Documentation has been updated where needed (mostly replacing Azure with Entra).