Adding additional tasks for record loading
Adds a task for selecting specific records from the cache for editing.

Adds a task for bulk loading changed records from a directory (presumed to be a WMS pre-head type working copy of records repo) for making larger changes to records via other scripts etc.

Existing load task renamed to import (which makes more sense).

In time much of this will be refactored into a Repo class or similar to coordinate better.